requirements.txt
opencv-python>=4.8.0
deepface>=0.0.79
customtkinter>=5.2.0
spotipy>=2.23.0
pillow>=10.0.0
numpy>=1.24.0